[IMP] mrp_workorder: update the reserved lot on WO in multi step manufacturing
Before this, in multi step manufacturing if lot is assigned through picking
validation, Lot number was not displayed on workorder.
Technically, the field was lot_id from workorder which is related to quality
checks.

After this, Lot number will be displayed for component which is reserved via
picking validation.

TaskID - 2480775

closes odoo/enterprise#18086

X-original-commit: 484d9fbe866b6aa8e8f7ab25d53efaac3c2e5fc5
Related: odoo/odoo#70207
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>
Signed-off-by: Rémy Voet <ryv-odoo@users.noreply.github.com>