[IMP] industry_fsm_sale: compute partner_id only for basic projects
Before this commit, the partner_id can be trigered for the fsm projects
when the user add a new employee mapping into these projects.

This commit filters the projects in the _compute_partner_id to only
continue the process for the basic projects.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164

closes #16356