[IMP] industry_fsm_sale: prevent useless recompute of price unit
Before this commit, when the user wants to change the price unit of an
employee mapping in a fsm project, the `_compute_price_unit` is triggered
because when the user has made a changes in any employee mapping of a
project, the `_compute_partner_id` of the project is trigger to find a
partner for the project based on the sale_line_id and the sale_line_id
of its mappings. After the `_compute_partner_id` is finished its process
the _compute_sale_line_id of employee mappings in this project is
triggered because the partner of the project has perhaps changed in its
compute. Finally, since the _compute_sale_line_id of employee mappings
is trigger then the _compute_price_unit is triggered to find the correct
price_unit and this compute removes the edition of the user.

This commit prevents to recompute the price_unit of employee mappings
when the _compute_sale_line_id is triggered and their project is a fsm
project. Because in the fsm projects, the SOL in employee mapping is not
displayed and not set.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164

closes #16356