[FIX] data_merge: duplicate rule "name (copy)"
Since the name field of data_merge.model is required and copy=False,
duplicating it gives you an error:

"""
The operation cannot be completed:
- Create/update: a mandatory field is not set.
- Delete: another model requires the record being deleted. If possible, archive it instead.
"""

That is not very friendly since the name is removed on copy.

With this changeset, we add ` (copy)` next to the previous name so it is
still possible to duplicate.

opw-2490990

closes odoo/enterprise#17611

X-original-commit: 0cc0ec4336caca743b57d659b3f790df6ceab28e
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>