[FIX] test_l10n_be_hr_payroll_account: wait for home page
The last step of `hr_contract_salary_tour_2` test tour navigates to
`/web/` but does not wait for the page to be fully loaded.

In rare cases, it leads to a race condition and makes the test fails
(see build error ID 1386 on runbot).

The race condition happens when the ChromeBrowser clear method is called
while the `web` page starts to load.

closes odoo/enterprise#17631

X-original-commit: 03c6e251fe3679891dc5f2f74403fc22da638923
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>