[IMP] planning: add notifications in gantt view
Before this commit, when the user is in the gantt view and want to copy
the planning of the last week, he can click on the button but he does
not know if the action is done or not, if there are some slots to copy
or not, etc. Moreover, when the user wants to send the planning, he does
not have any information to know if the email is sent or not for his
employees.

This commit adds a success notification after sending the schedule to
the employees of the user and it adds a success notification when we
found some slots to copy and a red (type='danger') notification when
there are no slots to copy.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958