[IMP] planning: display form view dialog when the employee needs email
Before this commit, when the user wants to send the schedule to his
employees and one of them has no email set in work_email field, the user
cannot know if an employee has no email set in this field. The system
ignores these employees and send the schedule for the employees whose
have an email address.

This commit checks if there is an employee with no email address in
work_email field, if it is the case then a form view dialog of
hr.employee.public is displayed to allow the user to add an email
address.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958