[FIX] sign: Display error with too many fields
Add a seperate scrollbar to the sign field type toolbar to prevent its content overflows the element's box when there are too many fields.

Before this commit was applied, when a document or a template is opened or uploaded, and if the user adds many fields to the left panel,
it will create a scrollbar on the right of the screen. This scrollbar allowed the user to view all fields.
But, the main window with the PDF content would be truncated and only visible with the help of a second "inner" scrollbar.

taskid:2501785

closes odoo/enterprise#17638

Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>