[FIX] account_reports: gather available tax reports properly in multicompany environment
account.generic.tax.report's _get_options() removes the multicompany key from the options if a config parameter is not set (by default, tax report is monocompany ; this parameter is only there to support VAT units). It's done after calling super(), so the filters are initialized with the multicompany key (by the super call); we can't rely on it in them.

closes odoo/enterprise#17681

X-original-commit: 85a4881b9a4a82659f31496dd3177dda7f52bce1
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>