[FIX] l10n_mx_edi: payment cfdi currency rate fixes
Before, the currency rate was the wrong one and was
not rounded and if we had different currencies the amount
was 0.

Now, we calculate the currency rate based on the partial
reconciliations to be as precise as possible.  We round
the rate to 6 numbers.  It needs to be how much invoice
currency you pay for one payment currency.  And we always
put the amounts.

A test was added as well.

For the moment, we have not implemented write-off yet.
(because it is a questionably legal corner case)

closes odoo/enterprise#17922

X-original-commit: 972f713aa4ec9e2077743c71ea79877f64b01e3a
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Josse Colpaert <jco@openerp.com>