[FIX] planning: fix cyclic dependency
This commit fixes a cyclic dependency between allocated_hours and
allocated_percentage. If both have to be recomputed, we make sur the
percentage doesn't change.

=== Allocated Hours and Percentage

The cyclic dependency between Allocated Hours and Percentage has been broken.

Previous behaviour:
When both allocated_percentage and allocated_hours had to be recomputed, the
order used to process field lead to different behaviours. On some corner
cases, i.e. with a 'forecast' slot, planned on multiple days, and during
the days it was planned on the employee had a time off, if you move the slot
to some days where there are no time off, the result will depend on the field
recomputed.
If it's the percentage, the percentage will increase and then influence the
allocated time. If it's the allocated time, the percentage won't be recomputed
so only the allocated time is updated.

Current behaviour:
Only the allocated time is recomputed if both fields have to be recomputed.

PR : odoo#15921

task-2215216