[IMP] account_reports: allow the use of new formulas with a groupby
The purpose is to overcome a limitation in the Financial Reports when accounts
position depend on their balance.
If globally the sum is positive, it does not take into account the individual balances.
The Financial Reports are currently limitative and do not allow to split balances
that are positive and negative.
The computation is made on the domain globally.

Task: 2502337