[IMP] social_youtube: update IAP routes to match new best practices
This should normally not be a change for stable but since the social_youtube
module is brand new and not used yet, we take this opportunity to change some
IAP routes to match the new naming best practices.
(All routes should contain a version number).

This will avoid having some deprecated routes right at the release of the
module.

See IAP PR for more information: odoo/iap-apps#284
Task-2209729

closes odoo/enterprise#17822

X-original-commit: d93e2b226747b271b6c0233fb20e26b03816e781
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>