[FIX] helpdesk{_fsm},industry_fsm{_sale}: update documentations
Before this commit, the links to the documentation in Helpdesk
and Field Service is for the v.13 and not v.14. Moreover, some
documentations in v.14 have been removed.

This commit removes the links in which the documentation has been
removed or will be removed and update the link to have the most recent
documentation.

closes odoo/enterprise#18015

X-original-commit: 472022ead942d75dce052d1afac08c873d6d912e
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>