[FIX] web_mobile: race condition in field many2one barcode tests
Then KanbanView in mobile has an additional JS lib dependency, which
means that it may needs to perform a request before it opens up.
However, a test in the field many2one barcode (which opens a dialog
containing a kanban view) was only waiting for a next tick (through
`dom.click()`), which is possibly too short for a network request.

So, depending on the network speed (and on the test order), this test
could fail. To fix it, we simply make sure that the test also wait for
the library to be loaded.

Based on commit odoo/odoo@02a75a560ca3b5866728b6405643500d8786697c

Also revert changes made in this file by commit odoo/enterprise@4a84fbb5c8aae1296d485488bdaeac94e00fa896
as they doesn't fix it and aren't useful.

closes odoo/enterprise#17725

X-original-commit: 7ede21aeef034bee37ebd5758dd3273e493ec0d1
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>