[IMP] planning: display notification when the shift is saved as template
Before this commit, when the user wants to save as template a shift in
the form view of it, if he checks the 'Save As Template' checkbox and
click on 'Save', he does not know if the shift is saved as template.

This commit displays a notification to inform the user that the shift is
saved as template.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958