[IMP] planning: display employee form view if no email
Before this commit, when the user is in a slot form view and wants to
send the slot to employee(s). If these employees can have no work email
set then the slot will not be send to employees.
Moreover when the user send planning to his employees, if an employee
has no email, the employee will not have his planning.

This commit displays an employee form view simplified to allow the user
to set a work email to his employee. If the user cancels the dialog form
view of the employee then we cancel the initial action.

task-2463722

Upgrade PR: odoo/upgrade#2426

closes #17958