[IMP] mrp_plm: allow updating the bom to latest version on MO
On a draft MO, if the bom is not the latest version, show a button that
can change it to the latest version. When the MO is already confirmed,
the button will show a message to alert the user a newer version is
available.

Task-2466232
COM PR odoo/odoo#69914
PR #17960