[IMP] hr_payroll,l10n_be_hr_payroll,hr_payroll_account_sepa: remake salary attachments
Completely reworks salary attachments, main points:
-> remove l10n_be_hr_payroll's version of salary attachments
-> add salary_attachments to hr_payroll
-> refactor payslip 'paid' state into hr_payroll instead of
payroll_account_sepa
-> salary attachments are now linked to the employee and not the
contract

Because salary attachments are not exclusive to belgium they have been
moved to the main payroll app. And they have been unlinked from the
contract to be linked to the employee instead, this is to prevent a
salary attachment from being removed when a contract gets voided
(cancelled/archived) and also to get a better overview of an employee's
attachments at once.

They have also been reworked to allow multiple things:
-> no end date required (in the case of child support for example)
-> a total amount can be set which will set the attachment as completed
if reached
-> attach a document
-> record payments

Tests have also been added for salary attachments.

The paid state has been refactored into the main payroll app because it
was used even without the sepa app and to allow attachment payments to
record automatically when a payslip is changed to the paid state.
This also means that the payslip flow has been changed a little:
draft->done->paid(->cancelled)
And thus a new button will appear "Mark as paid" when the state is done.
This button will be replaced by the sepa app as not to change the
previous behaviour when both apps are installed.
The same new behaviour is also present in payslip batches.
Flow tests have been modified for this new flow.

Task ID: 2481293