[FIX,ADD] l10n_be_hr_payroll_eco_voucher: Allow computing eco vouchers amount
In June of each year, eco-vouchers are paid to employees reporting to the Joint
Committee 218 Eco-vouchers are paid as long as this benefit has not been converted
into another benefit before 10/31 of the year who is before.

The eco-check constitutes an advantage intended for the acquisition of products and
services of an ecological nature included in a list set out in the appendix to CCT 98
concluded at the level of the National Labor Council. Workers can only acquire with
eco-vouchers those products or services of an ecological nature expressly mentioned
in this list. Their validity is limited to 24 months from the date they are made
available to the worker. The maximum face value of the eco-check is EUR 10 per eco-check.

closes odoo/enterprise#17864

Taskid: 2508613
X-original-commit: 8178c85e6e913492b6747aa8f758a669202ab604
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>