[FIX] account_reports: tax report: don't display fiscal position option on generic report
The generic tax report should only use 'all' as fiscal position option, and never allow selecting anything else. It used to allow selecting the fiscal positions of the domestic country, which was wrong.

[FIX] account reports: tax report: don't always keep 'all' fiscal position filter when coming from the generic report

The issue wasn't visible in the interface, but the options ended up being inconsistent doing this. To reproduce:

1) open the localized tax report from your fiscal country >> the fiscal position is 'domestic'

2) switch to the generic layout >> fiscal position option becomes 'all'

3) switch back to the localized report >> 'all' was kept, while 'domestic' should have been set instead

closes odoo/enterprise#17795

X-original-commit: 486c0b2b82008fb49373edde05e332e9a5f030cc
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>