[FIX] industry_fsm_sale: fix the compute_partner_id in project
Before this commit, the compute_partner_id should take only the
basic projects and not the fsm one, but it is not the case only the fsm
projects are used in this compute.

This commit corrects the compute to use only the basic projects as we
expect. With this changed, in strange behaviour reappears, that is the
user cannot manually change the price_unit field in employee mapping for
fsm projects because of the currency_id field needs to be recomputed and
then also recompute the price_unit field since the both fields are the
same compute method. This compute overrides the new compute method for
the currency_id field to not unnecessarily recompute the price_unit and
allow the user to manually edit the price unit as before.

Bug since this commit: 45136b3e380

Related PR: odoo/enterprise#16356

Community PR: odoo/odoo#69458

closes odoo/enterprise#17770

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>