[IMP] documents_spreadsheet: do not merge top-left cells on pivot insertion
Before this commit, the top left cells of a pivot are merged when
inserting in spreadsheet.

However, this merge can quickly become an issue when trying to select a
range. Since the initial motivation for merging these cells was purely
aesthetic, we simply get rid of this merge at pivot insertion.

Task-id 2510104

closes odoo/enterprise#17772

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>