[IMP] hr_payroll: Deprecate _get_salary_line_total for _get_line_values
Purpose
=======

A huge performace increase is expected by using this new method properly.

Imagine you need to retrieve the aggregation of 5 paylips lines for a
given set of payslips, you will make a single SQL request, instead of
looping 5 times on all the payslips lines to sum the values 1 by 1.