[FIX] helpdesk: fix traceback on helpdesk team creation
Currently, when we delete the 'customer care' team generated
by default and create a new team then it generates an error.
due to m2m deos not support the 0,0 create on as default value.
On previous version error was prevented because stage_ids field
is not loaded in the form view.

so in this commit, we fix this issue by creating a new stage_id if
the team doesn't get a default stage_id.

closes odoo/enterprise#18255

Taskid: 2507832
X-original-commit: 9c20c609a5a3158a9eb96349237ffff577420df8
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>