[FIX] mrp_plm: hide statusbar for newly created new eco
When creating a new ECO, the statusbar is always clickable before we
save it, regardless if you should have approvals in this stage or not.
Currently, we can't make the statusbar readonly or not clickable at one
specific state. Instead, we make the statusbar invisible before "Start
new revision".

Task-2466232
COM PR odoo/odoo#69914
PR #17960