[FIX] web_enterprise: fix home menu keyboard navigation issue
Before this commit: when user search in home menu and press right arrow key
and then press left arrow key then focus does not go to previous app icon,
it is because focus remain on search input box due to which issue occurs.

After this commit: when keyboard navigation done set focus to o_focused
element that is currently selected element on patched method, with this
commit when doing keyboard navigation it works smoothly with Right/Left
arrow key.

task-2508031

closes odoo/enterprise#19306

X-original-commit: 82f2ac7ac0abc8bd69cd7c7d2a46a9948b3df67f
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>