[IMP] l10n_pe_edi: improve how we handle the attachments for mail
There is a zip with the 2 XML we need so we are unzipping them and
inside the CDR there is a response that contains a base64 zip,
so, we are decoding-unziping-reading that data and attaching it to the
email we are going to send as XML.

The logic is as follows: datas(zip(xml(base64(zip(xml))))) there is no
method that can handle this. So, we are making a specific method for
this and adding a new one that will get the attachments inside of a
given base64 zipfile like data.  A method in the mail generation was
refactored so we can pass those 2 xmls instead of the original zip
in the mail.

closes odoo/enterprise#22568

X-original-commit: d7dcc182eb31759031fb954b612d6fcf9fb10981
Related: odoo/odoo#80490
Signed-off-by: Laurent Smet <las@openerp.com>