[FIX] account_reports: patch non-deterministic error on runbot when running the tests
The payment method selected by default by account.payment.register on nightly builds sometimes was SEPA Credit Transfer instead of 'manual', which crashed the tests as we didn't expect that, and the bank account set on the journal was not an IBAN, hence not validating the SEPA constraint.

We couldn't determine precisely the cause of this issue. This is a first attempt at fixing the test, it could require further investigation.

closes odoo/enterprise#17848

X-original-commit: cf967f7f95a11c69e4385fa9d84db93aa638268f
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>