[IMP] sign: reassign signatory
configure in sign roles to enable Change Authorized
enable reassign signatory in the form view of a sign request

Before, if the person who have to sign a document is sick or go away, we are blocked, nobody can sign the document.
This feature is really usefull when the document have to be by "Someone from HR Officer Department", or something else.
So the aims of this feature is to give the possibility to change a signatory.

task-2510209

closes odoo/enterprise#18384

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>