[FIX] approvals: add a step to properly finish tour
The last step of the approvals tour clicks on the `submit` button but
does not wait for the result.

As a consequence, the python test is considered as finished but some
requests may still to occur, depending of the speed. For example, if the
chrome browser is quickly killed, before anything else happens, there
are no remaining requests. In that case, the last step can be considered
as useless as it does not trigger anything server side.

But when the browser is killed to late, there could be up to 5
remaining requests. Worst, in some rare cases, some of them are not
waited by the HttpCase and may interfere with another test.
It happens on runbot, makeing the portal tour fail:
https://runbot.odoo.com/runbot/build/7213084
But this particular case is part of another investigated issue.

closes odoo/enterprise#17874

X-original-commit: 82feb5099edf68252a43a88906d95a2d2daa8412
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>