[IMP] social,social_push_notifications: use CRON triggers when scheduling posts
This commit implements 2 main changes related to the usage of CRON triggers:

1. When scheduling a social.post, we create a CRON trigger so that the post is
published on the social media as close to the selected date as possible.

This is important for example when you schedule a post on the 05/05/2021 at
exactly 9 AM, where you advertise that online ticketing for your online event
is now open for sale.
You want this post to be published as close as possible to 9 AM, and not
whenever the CRON job decides to run after that time.

2. When posting push notifications "immediately", they are now sent
immediately.

Push notifications implementation is a bit special because we force the send
method to pass through the CRON job.
As opposed to other social media where the API calls are made immediately when
you press on the "Post" button.

This is done to avoid having the user wait for a long period of time as push
notifications can quickly generate numerous HTTP calls (one for each visitor
you are trying to reach).

Now, with the usage of CRON trigger, we can actually send the push
notifications very close to when you press the "Post" button, by creating a
trigger "at=now()".

This also allows us to remove the overrides/workarounds done in
"website_event_social" that were used to send the track reminders "on time"
with manual buttons such as "Send Push Reminders" on the event and the
"action_post_immediate" method on the social.post that messed with the context.

Task-2336219
UPG PR odoo/upgrade#2417

closes odoo/enterprise#17852

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>