[FIX] taxcloud: Avoid unnecessary tax recomputations
A customer experienced 30-50 second delays after confirming the address
in the e-commerce shopping cart. The reason this happened is that some
of the Taxcloud code queries the first tax that meets certain criteria
and sets the `active` field to `True`. The ORM currently doesn't check
if this value was already the same, and will marked all fields and
records dependent on the tax for recomputation. In the case of the
customer that involved thousands of records and a dozen fields of those
records, which takes a while to recompute.

This fix avoids that scenario by only setting the value if it wasn't
already set.

closes odoo/enterprise#17942

Ticket: 2451513
X-original-commit: d5270048cda28737309d25fb055a7f13d523bc9e
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Raf Geens <raf-odoo@users.noreply.github.com>