[IMP] timesheet_grid: display an overtime indication in grid view
Before this commit, when the grid view of timesheets is grouped by
employee_id, we display the working schedule of each employee based of
his resource calendar. However the manager should know the working
schedule of his employees, but if he has a good indication of the
workload of each employee, this can help the manager to find a better
balance his resources accross projects.

This commit changes the working schedules for each employee by the
difference between the number of hours the employee is supposed to work
on that period according to his resource calendar and the duration of
his timesheets during that same period.

task-2500800

closes odoo/enterprise#17959

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>