[FIX] hr_payroll: fix defer time
A time off may be blocked from being approved if two payslips for that
time period exist if one of them is in a done state and the other is in
draft.
As we still want to be able to accept the time off in this case, we now
check for draft payslips aswell for that time period.

Added some tests to keep this behaviour.

Task ID: 2486703

closes odoo/enterprise#17280

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>