[FIX] amazon: fix concurrent access error prevent SO to create
Any concurrent access error during the creation of a SO in amazon's app will prevent the SO to create. As the last_sync_date will still update, this order won't be sync except if you backtrack the date manually.

By raising an exception of the `PG_CONCURRENCY_ERRORS_TO_RETRY` kind,
we let the default Odoo's behavior happen, which is:
 - If it is in an action (button "Sync order" in Amazon's setting, manually run the schedule action), it will retry the request. The entire request cursor is rollback then re-run after some delay.
 - It it is the scheduled action, it will raise the error (with a traceback) and rollback the current cursor
 Technical note: When the cursor do a rollback, it does discard what wasn't committed yet. As Amazon's module does commit the cursor frequently, it shouldn't have a big impact on Amazon's API calls.

On the ticket 2488349, the concurrent access happened because the "Inter-Company Transactions" was enabled. As such, PO were created/edited while the SO cursor was closing. If anything during this process edit the PO line a concurrency will happen. This was the case as a (custom) Scheduled action was editing PO lines. As they are on odoo.sh, sometime this scheduled action ran alongside the amazon's one.
OPW-2488349

closes odoo/enterprise#18110

X-original-commit: 550b86ccf2af1f57bc1018dbcf689e2aeca0224c
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>