[FIX] web_enterprise: avoid calendar error in mobile month view
This PR fix a bug introduced in odoo/odoo@ac13964244f54f268635b3a5d209d0b8e7d78076
Before this commit, on mobile, 'this' was undefined because
'_getFullCalendarOptions' was overridden and so the scope was not
properly passed along.

After this commit, now we bind the event with the FullCalendar
instance to keep the original 'this' through the override.

Steps to reproduce:
* Open Odoo (Mobile)
* Go in a calendar view
* Select month view (with a event inside) => BUG

closes odoo/enterprise#17931

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>