[FIX] sale_subscription_dashboard: add sign when computing MRR
The computation of the monthly recurring revenue does not consider the
difference between an invoice and a refund.

To reproduce the error:
1. Create a SO with a monthly subscription
2. Confirm
3. Create Invoice, Post it
4. Add Credit Note, Post it
5. Subscriptions > Reporting > Revenue KPIs
    - Include current date in dates interval

Error: The value of MRR is $400. The module summed up the invoice and
the credit note without considering that the sign of the note should be
negative.

OPW-2487560

closes odoo/enterprise#18119

X-original-commit: 41cc728128468159ead3503a6b32a61210cb0586
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>