[FIX] web_enterprise: convert home menu to list view for iOS app
According to the app-store guidelines we are not allowed to display
a "store-like interface":
"2.5.8 Apps that create alternate desktop/home screen environments
or simulate multi-app widget experiences will be rejected."

Our updates are currently blocked because of this.
In the rush, we have no choice but to change the "app switcher" interface
although this will confuse users for sure.
It's why we only do this change for iOS users using the mobile app.
For the moment, we have been granted an exception for iPads.

Source: https://developer.apple.com/app-store/review/guidelines/#software-requirements

Related PR:
https://github.com/odoo/mobile-ios/pull/104

Task-id: 2483253
X-original-commit: 85e379f3ad8f56361f6eda7e7adce4ccdc94f61a