[IMP] industry_fsm_{sale,stock}: update delivered quantity when task done
Before this commit, when the Inventory app is installed, the DO linked
to the SO is automatically validated once the task is marked as done.
However, when this app is not installed and some materials are added to
the task, these materials are not automatically delivered once the task
is marked as done and thus cannot be invoiced. This is frustrating
because it can block the creation of the invoice, and it also gives the
impression that the invoice was not correctly created.

This commit updates the delivered quantity of the consumables added to
the task when this task is marked as done.

task-2501713

closes odoo/enterprise#18101

Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>