[FIX] industry_fsm_sale: compute bad amount in so when task mark as done
Before this commit, when the user who has no access right in Sales app
and mark as done a task with some materials, when we generate new SOLs
for timesheets in the task, the total amount of the SO is recompute and
if we do not add sudo then we compute the total amount of SO based on
the SOLs that we have generated and not all SOLs in the SO.

This commit changes the sudo(False) in sudo() when we prefetch the SOLs
of the SO to prefetch and keep the sudo in all SOLs. Thus, the recompute
of total amount will be good.

Step to reproduce the bug:
1) Login in Odoo with an user who has no access right in Sales and
Accounting apps
2) Go to Field Services App
3) Create a fsm task
4) Define a partner in this task and add some materials
5) Return on the task and click on "mark as done" button
6) Logout and login with an user who can see the SO generated with this
task and you will see the total amount is not the one expected.

Related PR: #16348

closes odoo/enterprise#18020

X-original-commit: 917c7a002f0ee40af33c4251b8d9fe03dee6f8d7
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>