[IMP] various: clean duration in days
helpdek_timesheet, industry_fsm(_report),
project_timesheet_forecast(_sale)

The purpose of this task is there are several places where the
duration is expressed in hours even though the encoding unit
is in Days. This creates inconsistencies.

In this commit, we convert hours into days when the encoding
unit is in Days. Also, we changed the string of the fields
according to the encoding unit.

Related Community PR: https://github.com/odoo/odoo/pull/70512

closes odoo/enterprise#18197

Task-id: 2512950
Related: odoo/odoo#70512
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>