[IMP] documents_spreadsheet: Date filters need a default year value
Currently, when creating a non-yearly Date filter (i.e. of type 'month' or 'quarter'),
the user will first select the period, which will then automatically
select the current year in the year selection field. This is actually
strictly visual, a result of rerendering of the owl component
`DateFilterValue`and at that point of time, the owl component in charge
of creating the filter doesn't have a clue of what year is selected
since the event in charge of transmitting this information was never
triggered. The user can then save his filter, ultimately craeting a new
filter with a given period but no year specified.

While it does work when updating the pivot domain (we just add
`year:"this_year"` if it's not defined) this proves to be problematic
when using the function FILTER.VALUE.

e.g. =FILTER.VALUE("my label without year")  => Q4 instead of Q4 2021