[FIX] web_studio: make test pass on chrome 90
The changed test uses the drag&drop helper, and an operation does
not work as expected with the given params on chrome 90. The runbot
currently uses chrome 80, so it is not an issue, but if your chrome
is up-to-date, and you try to run the test suite, this test would
fail.

Here, we want to drop a text span after another text span, i.e. on
its right. Before this commit, it was dropped on its exact center,
which appearently was considered as 'after' before chrome 90, but
'before' as of chrome 90. Specifying explicitely that we want to
drop it after is harmless and makes the test more robust.

closes odoo/enterprise#18060

X-original-commit: b01c4d3cbe38aec2ea3d8bf4c4690d5e7738f1b2
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>